Reorganize the lib/ folder
