+++ pages/partners.tsx
@@ -8,7 +8,7 @@ import { Layout, Section, SectionContent } from "components/layout";
 import * as S from "components/partners/styles";
 import Tabs from "components/tabs";
 import { siteData } from "lib/site-data";
-import { Article } from "lib/related-blog-posts";
+import { Article } from "lib/data-sources/blog";
 import { PortalPartner } from "lib/airtable/partner";
 
 type PageProps = {
