+++ pages/projects/[slug].tsx
@@ -12,7 +12,7 @@ import { ParsedUrlQuery } from "querystring";
 import { siteData } from "lib/site-data";
 import OpportunityItem from "components/sections/opportunity-overview";
 import { Route } from "lib/utils";
-import { Article } from "lib/related-blog-posts";
+import { Article } from "lib/data-sources/blog";
 import { BlogCard } from "components/cards";
 import EventCard from "components/dashboard/event-card";
 import { PortalUser } from "lib/airtable/user";
